refactor: extract metrics to resolve circular dependency and remove deprecated checks module
- Extracted `get_loc`, `get_complexity_score`, `check_type_hints`, `calculate_acl`, `count_tokens`, and `get_function_stats` to `src/agent_scorecard/metrics.py`.
- Updated `src/agent_scorecard/analyzer.py` to import from `.metrics` and re-export symbols for backward compatibility.
- Updated `src/agent_scorecard/scoring.py` to import from `.metrics`, resolving the circular dependency between `analyzer` and `scoring`.
- Deleted the deprecated `src/agent_scorecard/checks.py`.
- Verified that all tests pass and the CLI functions correctly.
